changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11845 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/IPE/IpeSimulation/ORIpeSimulationController.h
@@ -147,7 +147,7 @@
 #pragma mark ‚Ä¢‚Ä¢‚Ä¢Data Source Methods (OutlineView)
 - (int) outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item;
 - (BOOL) outlineView:(NSOutlineView *)outlineView isItemExpandable:(id)item; 
-- (id) outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item; 
+- (id) outlineView:(NSOutlineView *)outlineView child:(NSUInteger)index ofItem:(id)item; 
 - (id) outlineView:(NSOutlineView *)outlineView objectValueForTableColumn:(NSTableColumn *)tableColumn byItem:(id)item;
 
 #pragma mark ‚Ä¢‚Ä¢‚Ä¢Data Source Methods (ComboBox)
